Tried adding adjacent reference by array index
Running into issues with left/right being backwards. This is not just as easy as switching them because I don't understand why they are being drawn backwards so the conditionals moving forward would get messy if my understanding is unclear.

Top and bottom are also having issues, but these will be addressed after left and right should I choose to move forward with this method of reference.