Replaced rulesets with modular rules
instead of preset tick functions selected by a switch statement, each num of neighbors has its own stateX var assigned to it.

These will be associated with divs that will either make the cell die, live, both, or do nothing based on the number of neighbors. These divs will be toggled on and off with a mouse click.